Use 32 bit float precision

+++ main/include/nodes/thermal_camera.hpp
@@ -40,7 +40,7 @@ private:
     bool updateThermistor;
     std::vector<uint8_t> thermistorBuffer;
     std::vector<uint8_t> pixelBuffer;
-    double pixels[AMG88XX_PIXEL_ARRAY_SIZE];
+    float pixels[AMG88XX_PIXEL_ARRAY_SIZE];
     int32_t time;
 
     void updateTimerCallback(rcl_timer_t *timer, int64_t n);
